Add API get user by Id

+++ src/main/java/UserDomain/repository/DoctorRepository.java
@@ -3,5 +3,5 @@ package UserDomain.repository;
 import UserDomain.model.Doctor;
 import org.springframework.data.repository.CrudRepository;
 
-public interface DoctorRepository extends CrudRepository<Doctor, Integer> {
+public interface DoctorRepository extends CrudRepository<Doctor, Long> {
 }
